# app/src/docs/v1.api-spec.yaml
...
openapi: 3.0.2
info:
  version: 1.0.0
  title: NR PermitConnect Navigator Service (PCNS)
  description: >-
    A case management application meant to serve the needs of the NRM Permitting
    Solutions Branch.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
servers:
  - url: /api/v1
    description: This server
security:
  - BearerAuth: []
    OpenID: []
...
paths:
...
  /electrification/project:
    get:
      summary: Get a list of electrification projects
      operationId: getElectrificationProjects
      tags:
        - Electrification Project
      responses:
        '200':
          description: A list of electrification projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DB-ElectrificationProject'
        "401":
          $ref: "#/components/responses/Unauthorized"
        '403':
          $ref: '#/components/responses/Forbidden'
        default:
          $ref: '#/components/responses/Error'
    put:
      summary: Create a electrification project
      operationId: createElectrificationProject
      tags:
        - Electrification Project
      responses:
        '201':
          description: Electrification Project created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Response-ElectrificationProjectCreate'
        "401":
          $ref: "#/components/responses/Unauthorized"
        '403':
          $ref: '#/components/responses/Forbidden'
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        default:
          $ref: '#/components/responses/Error'
  /electrification/project/{electrificationProjectId}:
    get:
      summary: Get a specific electrification project
      operationId: getElectrificationProject
      tags:
        - Electrification Project
      parameters:
        - $ref: '#/components/parameters/Path-ElectrificationProjectId'
      responses:
        '200':
          description: Details of the specific electrification project
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DB-ElectrificationProject'
        "401":
          $ref: "#/components/responses/Unauthorized"
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        default:
          $ref: '#/components/responses/Error'
    put:
      summary: Update a electrification project
      operationId: updateElectrificationProject
      tags:
        - Electrification Project
      parameters:
        - $ref: '#/components/parameters/Path-ElectrificationProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Request-ElectrificationProjectUpdate'
      responses:
        '200':
          description: Electrification Project updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DB-ElectrificationProject'
        "401":
          $ref: "#/components/responses/Unauthorized"
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        default:
          $ref: '#/components/responses/Error'
    delete:
      summary: Delete an electrification project
      description: Delete of an electrification project
      operationId: deleteElectrificationProject
      tags:
        - Electrification Project
      parameters:
        - $ref: '#/components/parameters/Path-ElectrificationProjectId'
      responses:
        '204':
          description: Returns no content success
        "401":
          $ref: "#/components/responses/Unauthorized"
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        default:
          $ref: '#/components/responses/Error'
...
  /electrification/project/search:
    get:
      summary: Search electrification projects
      description: Returns a list of electrification projects based on the given queries
      operationId: searchElectrificationProjects
      tags:
        - Electrification Project
      parameters:
        - $ref: '#/components/parameters/Query-IncludeUser'
        - in: query
          name: activityId
          required: false
          schema:
            type: array
            description: The Activity IDs to filter on
            items:
              $ref: '#/components/parameters/Path-ActivityId'
        - in: query
          name: createdBy
          required: false
          schema:
            type: array
            description: The User IDs to filter on
            items:
              $ref: '#/components/parameters/Query-CreatedBy'
        - in: query
          name: electrificationProjectId
          required: false
          schema:
            type: array
            description: The Electrification Project IDs to filter on
            items:
              $ref: '#/components/parameters/Path-ElectrificationProjectId'
      responses:
        '200':
          description: A list of electrification projects matching the search criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DB-ElectrificationProject'
        "401":
          $ref: "#/components/responses/Unauthorized"
        '403':
          $ref: '#/components/responses/Forbidden'
        "422":
          $ref: "#/components/responses/UnprocessableEntity"
        default:
          $ref: '#/components/responses/Error'
...
components:
  parameters:
...
    Path-ActivityId:
      in: path
      name: activityId
      description: ID of an Activity
      required: true
      schema:
        type: string
        example: 2DE67F13
...
    Path-ElectrificationProjectId:
      in: path
      name: electrificationProjectId
      description: UUID of a Electrification Project
      required: true
      schema:
        type: string
        format: uuid
        example: da5c5031-0e84-4234-7766-1ba846997482
...
    Query-CreatedBy:
      in: query
      name: createdBy
      description: UUID of user who created data
      required: false
      schema:
        type: string
        format: uuid
        example: b76469a3-896c-4435-aa83-b8b87db6f701
    Query-IncludeUser:
      in: query
      name: includeUser
      description: >-
        Whether to include User data with each entity being returned by the results.
      required: false
      schema:
        type: boolean
        example: true
...
  schemas:
...
    DB-ElectrificationProject:
      type: object
      required:
        - electrificationProjectId
        - activityId
        - submittedAt
        - aaiUpdated
      properties:
        electrificationProjectId:
          type: string
          description: UUID for the electrification project
          format: uuid
          example: d95f1de6-698b-4d9d-b938-487eb446ace8
        activityId:
          type: string
          description: Activity ID
          example: D95F1DE6
        assignedUserId:
          type: string
          description: User assigned to the electrification project
          example: Smith, John
        submittedAt:
          type: string
          description: Date it was submitted
          format: date-time
          example: '2024-04-03T00:57:09.070Z'
        intakeStatus:
          type: string
          description: Intake status type
          example: Submitted
        applicationStatus:
          type: string
          description: Status of the application
          example: New
        queuePriority:
          type: number
          description: Navigator assigned priority rank
          example: 3
        projectName:
          type: string
          description: Name of the Project
          example: New Project
        projectDescription:
          type: string
          description: A description of the project
          example: Project description example.
        companyNameRegistered:
          type: string
          description: Name of the company connected to the project
          example: QW URBAN FOOD LTD.
        projectType:
          type: string
          description: The type of project
          example: IPP Wind
        bcHydroNumber:
          type: string
          description: Assigned BC Hydro number
          example: 1234ABCD
        submissionType:
          type: string
          description: The type of submission
          example: Assistance
        hasEpa:
          type: string
          description: Indicates if an EPA document is present
          example: 'yes'
        bcEnvironmentAssessNeeded:
          type: string
          description: Indicates if BC Environment Assessment is needed
          example: 'no'
        megawatts:
          type: number
          description: Project capacity in megawatts
          example: 5.5
        locationDescription:
          type: string
          description: Description of the project location
          example: Near the main substation
        aaiUpdated:
          type: boolean
          description: Flag indicating if AAI was updated
          example: false
        atsClientId:
          type: integer
          description: ATS client identifier
          example: 789078
        astNotes:
          type: string
          description: AST notes
          example: Initial notes
...
    Request-ElectrificationProjectUpdate:
      title: Request to Update Electrification Project
      type: object
      required:
        - housingProjectId
        - activityId
        - submittedAt
      properties:
        electrificationProjectId:
          type: string
          description: UUID for the electrification project
          format: uuid
          example: d95f1de6-698b-4d9d-b938-487eb446ace8
        activityId:
          type: string
          description: Activity ID
          example: D95F1DE6
        submittedAt:
          type: string
          description: Date it was submitted
          format: date-time
          example: '2024-04-03T00:57:09.070Z'
...
    Response-Error:
      title: Response Error
      type: object
      allOf:
        - $ref: '#/components/schemas/Response-Problem'
        - type: object
          properties:
            status:
              example: 500
            title:
              example: Internal Server Error
            type:
              example: https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/500
    Response-Forbidden:
      title: Response Forbidden
      type: object
      allOf:
        - $ref: '#/components/schemas/Response-Problem'
        - type: object
          properties:
            type:
              example: https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/403
            title:
              example: Forbidden
            status:
              example: 403
            detail:
              example: User lacks permission to complete this action
    Response-NotFound:
      title: Response Not Found
      type: object
      allOf:
        - $ref: '#/components/schemas/Response-Problem'
        - type: object
          properties:
            status:
              example: 404
            title:
              example: Not Found
            type:
              example: https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/404
    Response-Problem:
      title: Response Problem
      type: object
      required:
        - type
        - title
        - status
      properties:
        type:
          type: string
          description: What type of problem, link to explanation of problem
        title:
          type: string
          description: Title of problem, generally the HTTP Status Code description
        status:
          type: string
          description: The HTTP Status code
        detail:
          type: string
          description: >-
            A short, human-readable explanation specific to this occurrence of
            the problem
        instance:
          type: string
          description: >-
            A URI reference that identifies the specific occurrence of the
            problem
...
    Response-ElectrificationProjectCreate:
      title: Response for Create Electrification Project
      type: object
      required:
        - activityId
        - housingProjectId
      properties:
        activityId:
          type: string
          description: Activity ID
          example: D95F1DE6
        housingProjectId:
          type: string
          description: UUID for the housing project
          format: uuid
          example: d95f1de6-698b-4d9d-b938-487eb446ace8
...
    Response-Unauthorized:
      title: Response Unauthorized
      type: object
      allOf:
        - $ref: '#/components/schemas/Response-Problem'
        - type: object
          properties:
            type:
              example: https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/401
            title:
              example: Unauthorized
            status:
              example: 401
            detail:
              example: Invalid authorization credentials
    Response-ValidationError:
      title: Response Validation Error
      type: object
      allOf:
        - $ref: "#/components/schemas/Response-Problem"
        - type: object
          required:
            - errors
          properties:
            errors:
              type: array
              items:
                type: object
                required:
                  - message
                properties:
                  value:
                    type: object
                    description: Contents of the field that was in error.
                    example: {}
                  message:
                    type: string
                    description: The error message for the field.
                    example: Invalid value `encoding`.
            status:
              example: 422
            title:
              example: Unprocessable Entity
            type:
              example: https://developer.mozilla.org/en-US/docs/Web/HTTP/Status/422
  responses:
...
    Error:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response-Error'
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response-Forbidden'
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response-NotFound'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Response-Unauthorized'
    UnprocessableEntity:
      description: Unprocessable Content (Generally validation error(s))
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Response-ValidationError"
  securitySchemes:
    BearerAuth:
      type: http
      description: Bearer token auth using an OIDC issued JWT token
      scheme: bearer
      bearerFormat: JWT
    OpenID:
      type: openIdConnect
      description: OpenID Connect endpoint for acquiring JWT tokens
      openIdConnectUrl: >-
        https://logonproxy.gov.bc.ca/auth/realms/your-realm-name/.well-known/openid-configuration